% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{flowFrameFromFlowJoCSV}
\alias{flowFrameFromFlowJoCSV}
\alias{flowFrameFromFlowJoCSV,insideFlow-method}
\title{Build a Flowframe from CSV file expression matrix}
\usage{
flowFrameFromFlowJoCSV(object, dropCompPrefix = FALSE)

\S4method{flowFrameFromFlowJoCSV}{insideFlow}(object, dropCompPrefix = FALSE)
}
\arguments{
\item{object}{A insideFlow object}

\item{dropCompPrefix}{Boolean to allow for paramName cleanup}
}
\value{
Returns reversed transformed data matrix before export to FlowJo
}
\description{
Build a Flowframe from CSV file expression matrix
}
\examples{
\dontrun{
myobject <- flowFrameFromFlowJoCSV(object, popName, coFact, outsFolder)}
}
