% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{getParNames}
\alias{getParNames}
\alias{getParNames,insideFlow-method}
\title{View parameter names}
\usage{
getParNames(object)

\S4method{getParNames}{insideFlow}(object)
}
\arguments{
\item{object}{A insideFlow object}
}
\value{
Returns the parameter names from the exported population
}
\description{
View parameter names
}
\examples{
\dontrun{
getParNames(object)}
}
