% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{getRunID}
\alias{getRunID}
\alias{getRunID,insideFlow-method}
\title{View the unique run identifier}
\usage{
getRunID(object)

\S4method{getRunID}{insideFlow}(object)
}
\arguments{
\item{object}{A insideFlow object}
}
\value{
Returns the runID generated from the plugin process.
}
\description{
View the unique run identifier
}
\examples{
\dontrun{
getRunID(object)}
}
