% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{getStainNames}
\alias{getStainNames}
\alias{getStainNames,insideFlow-method}
\title{View stain names}
\usage{
getStainNames(object)

\S4method{getStainNames}{insideFlow}(object)
}
\arguments{
\item{object}{A insideFlow object}
}
\value{
Returns the stain names if present from the exported population
}
\description{
View stain names
}
\examples{
getStainNames(object)
}
