% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{class}
\name{insideFlow-class}
\alias{insideFlow-class}
\alias{insideFlow}
\title{The InsideFlow Class}
\description{
The insideFlow object is a representation of a gated population of cells that
are exported from a FlowJo plugin process. These files loaded into the object
can be either .fcs files or .csv files containing expression data from a flow
cytometery experiment. The object stores metadata as well as representations
of low-dimensional data from clustering or dimensionality reductions methods.

Each insideFlow object has a number of slots which store information. Key
slots to access are listed below.
}
\section{Slots}{

\describe{
\item{\code{runID}}{The unique id for plugin process}

\item{\code{inputCSV}}{The expression matrix exported from FlowJo}

\item{\code{parameterNames}}{The metadata for parameter and stain names}

\item{\code{flowFrameFJ}}{The expression matrix stored as flowFrame}

\item{\code{BatchCorrect}}{The expression matrix both uncorrected and batch corrected}

\item{\code{PeacoQCPops}}{The results from performing QC on flow data with PeacoQC}

\item{\code{FlowAIPops}}{The results from performing QC on flow data with FlowAI}

\item{\code{Rphenograph}}{The results from clustering with Phenograph}

\item{\code{FastPhenograph}}{The results from the FastPG algorithm}

\item{\code{FlowSOMClusters}}{The results from FlowSOM clustering}

\item{\code{DimReduction}}{The low-dimensional embedding result and map}

\item{\code{ClustRCheck}}{The results from cluster evaluation with ClustRCheck}

\item{\code{Taylor}}{The results from cluster evaluation via Taylor Index}
}}

