% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{loadFlowJoCSV}
\alias{loadFlowJoCSV}
\alias{loadFlowJoCSV,insideFlow-method}
\title{Load data into insideFlow object}
\usage{
loadFlowJoCSV(object, csvPath, eventNumber = "Event #")

\S4method{loadFlowJoCSV}{insideFlow}(object, csvPath, eventNumber = "Event #")
}
\arguments{
\item{object}{A insideFlow object}

\item{csvPath}{The path to the exported csv file from FlowJo}

\item{eventNumber}{The column name for event numbers or cell ids}
}
\value{
Returns the newly created object
}
\description{
Load data into insideFlow object
}
\examples{
\dontrun{
inputCSVpath <- "/Documents/ExtNode.csv"
myobject <- loadFlowJoCSV(object, inputCSVpath, eventNumber = "Event #")}
}
