% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{normalizedPlots}
\alias{normalizedPlots}
\alias{normalizedPlots,insideFlow-method}
\title{Return plots comparing markers after batch normalization}
\usage{
normalizedPlots(object, fjPopName, fjCofactor, outPutFolder)

\S4method{normalizedPlots}{insideFlow}(object, fjPopName, fjCofactor, outPutFolder)
}
\arguments{
\item{object}{A insideFlow object}

\item{fjPopName}{The name of selected population in FlowJo}

\item{fjCofactor}{The selected Cofactor value to apply}

\item{outPutFolder}{The path to the output folder}
}
\value{
Returns the plots from cyCombine showing normalized histograms
}
\description{
Return plots comparing markers after batch normalization
}
\examples{
\dontrun{
normalizedPlots(object, popName, coFact, outsFolder)}
}
