% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{plotFlowSOMResults}
\alias{plotFlowSOMResults}
\alias{plotFlowSOMResults,insideFlow-method}
\title{Plot FlowSOM Images}
\usage{
plotFlowSOMResults(
  object,
  plotAllAsStars,
  plotPar,
  radarColor = "viridis",
  plotBgd = "magma",
  numClust
)

\S4method{plotFlowSOMResults}{insideFlow}(
  object,
  plotAllAsStars,
  plotPar,
  radarColor = "viridis",
  plotBgd = "magma",
  numClust
)
}
\arguments{
\item{object}{A insideFlow object}

\item{plotAllAsStars}{Boolean to plot MST by selected param or with piecharts}

\item{plotPar}{Selected parameter to plot as heatmapped MST}

\item{radarColor}{The selected color palette for minimum spanning tree plot}

\item{plotBgd}{Color palette for mst nodes}

\item{numClust}{Number of metaclusters requested by user}
}
\value{
Returns reversed transformed data matrix before export to FlowJo
}
\description{
Plot FlowSOM Images
}
\examples{
\dontrun{
myobject <- plotFlowSOMResults(object, popName, coFact, outsFolder)}
}
