% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{prepareBEC}
\alias{prepareBEC}
\alias{prepareBEC,insideFlow-method}
\title{Transform data before batch correction step}
\usage{
prepareBEC(object, fjMarkersList, fjCofactor)

\S4method{prepareBEC}{insideFlow}(object, fjMarkersList, fjCofactor)
}
\arguments{
\item{object}{A insideFlow object}

\item{fjMarkersList}{The selected list of markers to normalize}

\item{fjCofactor}{The selected Cofactor value to apply}
}
\value{
Returns the object with asinh-transformed values to batch correct
}
\description{
Transform data before batch correction step
}
\examples{
\dontrun{
myobject <- prepareBEC(object, markers, cofact)}
}
