% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{reverseTransform}
\alias{reverseTransform}
\title{Reverse transforms data matix before exporting to FlowJo}
\usage{
reverseTransform(object, fjCofactor)
}
\arguments{
\item{object}{A insideFlow object}

\item{fjCofactor}{The selected Cofactor value to apply}
}
\value{
Returns reversed transformed data matrix before export to FlowJo
}
\description{
Reverse transforms data matix before exporting to FlowJo
}
\examples{
\dontrun{
myobject <- reverseTransform(object, popName, coFact, outsFolder)}
}
