% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{runBatch_correct}
\alias{runBatch_correct}
\alias{runBatch_correct,insideFlow-method}
\title{Performs batch correction with cyCombine}
\usage{
runBatch_correct(object, batchID, normMethod, xdim, ydim)

\S4method{runBatch_correct}{insideFlow}(object, batchID, normMethod, xdim, ydim)
}
\arguments{
\item{object}{A insideFlow object}

\item{batchID}{The selected list of markers to normalize}

\item{normMethod}{The selected Cofactor value to apply}

\item{xdim}{The first dimension of SOM used}

\item{ydim}{The second dimension of SOM used}
}
\value{
Returns the batch corrected object
}
\description{
Performs batch correction with cyCombine
}
\examples{
\dontrun{
myobject <- runBatch_correct(object, bID, nrmMethod, xdim, ydim)}
}
