% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{cleanupFJobj}
\alias{cleanupFJobj}
\title{Clean up object before saving}
\usage{
cleanupFJobj(object)
}
\arguments{
\item{object}{A insideFlow object}
}
\value{
Returns the light-weight object cleaned of transformed data slots
from batch correction processes
}
\description{
Clean up object before saving
}
\examples{
myobject <- cleanupFJobj(object)
}
