% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\docType{methods}
\name{cleanupFJobj}
\alias{cleanupFJobj}
\alias{cleanupFJobj,insideFlow-method}
\title{Cleanup Object}
\usage{
cleanupFJobj(object)

\S4method{cleanupFJobj}{insideFlow}(object)
}
\arguments{
\item{object}{A insideFlow object}
}
\value{
Returns the light-weight object cleaned of transformed data slots
from batch correction processes
}
\description{
Cleanup Object
}
\examples{
\dontrun{
myobject <- cleanupFJobj(object)}
}
