% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{loadFlowJoCSV}
\alias{loadFlowJoCSV}
\title{Load data into insideFlow object}
\usage{
loadFlowJoCSV(object, csvPath)
}
\arguments{
\item{object}{A insideFlow object}

\item{csvPath}{The path to the exported csv file from FlowJo}
}
\value{
Returns the newly created object
}
\description{
Load data into insideFlow object
}
\examples{
\dontrun{
inputCSVpath <- "/Documents/ExtNode.csv"
myobject <- loadFlowJoCSV(object, inputCSVpath)}
}
