% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideFlow.R
\name{loadFlowJoDataTable}
\alias{loadFlowJoDataTable}
\title{Load a data.table into insideFlow object}
\usage{
loadFlowJoDataTable(object, data)
}
\arguments{
\item{object}{A insideFlow object}

\item{data}{The data object to load into insideFlow object}
}
\value{
Returns the newly created object
}
\description{
Load a data.table into insideFlow object
}
\examples{
exMatrix <- fread("/Documents/ExtNode.csv")
myobject <- loadFlowJoDataTable(object, exMatrix)
}
